Lock down to 0.13; Document FluxibleComponent; Deprecate APIs
