+++ lib/FluxibleContext.js
@@ -6,6 +6,8 @@
 
 var debug = require('debug')('Fluxible:Context');
 var objectAssign = require('object-assign');
+var React = require('react');
+var FluxibleComponent = React.createFactory(require('./FluxibleComponent'));
 require('setimmediate');
 
 /**
@@ -44,9 +46,9 @@ FluxContext.prototype.createElement = function createElement(props) {
     if (!Component) {
         throw new Error('A component was not specified.');
     }
-    return Component(objectAssign({}, props, {
+    return FluxibleComponent({
         context: this.getComponentContext()
-    }));
+    }, Component(props));
 };
 
 /**
@@ -83,10 +85,8 @@ FluxContext.prototype.executeAction = function executeAction(action, payload, do
     var self = this;
     payload = payload || {};
     var displayName = action.displayName || action.name;
-    setImmediate(function executeActionImmediate() {
-        debug('Executing action ' + displayName + ' with payload', payload);
-        action(self.getActionContext(), payload, done);
-    });
+    debug('Executing action ' + displayName + ' with payload', payload);
+    action(self.getActionContext(), payload, done);
 };
 
 /**
