+++ package.json
@@ -20,7 +20,7 @@
     "setimmediate": "^1.0.2"
   },
   "peerDependencies": {
-    "react": ">=0.12.0 <=0.13.x"
+    "react": "0.13.x"
   },
   "devDependencies": {
     "chai": "^2.0.0",
