+++ utils/createStore.js
@@ -1,4 +1,3 @@
-// @TODO deprecate
-// console.warn("require('fluxible/utils/createStore') is deprecated. Please use " +
-//"require('fluxible/addons').BaseStore or require('fluxible/addons/createStore').");
+console.warn("require('fluxible/utils/createStore') is deprecated. Please use " +
+"require('fluxible/addons').BaseStore or require('fluxible/addons/createStore').");
 module.exports = require('../addons/createStore');
